bai 13: understand middleware
