+++ controllers/user.controller.js
@@ -1,6 +1,7 @@
 const db = require('../db')
 const shortid = require('shortid')
 
+
 module.exports.index = (req, res) => {
     res.render('users/index', {
         users: db.get('users').value()
@@ -31,20 +32,6 @@ module.exports.get = (req, res) => {
 
 module.exports.postCreate =(req, res) => {
     req.body.id = shortid.generate();
-    let errors = []
-    if(!req.body.name){
-        errors.push('Name is required')
-    }
-    if(!req.body.phone){
-        errors.push('Phone is required')
-    }
-    if(errors.length){
-        res.render('users/create', {
-            errors : errors, 
-            values: req.body          
-        })
-        return ;
-    }
     db.get('users').push(req.body).write()
     res.redirect('/users')
 }
