+++ db.json
@@ -3,28 +3,6 @@
     {
       "name": "thong",
       "id": "scFhiMWfC"
-    },
-    {
-      "name": "teo",
-      "id": "zCYByiFpm"
-    },
-    {
-      "name": "david",
-      "id": "yoBMOVzuA"
-    },
-    {
-      "name": "ti",
-      "id": "POnd-kI1F"
-    },
-    {
-      "name": "minh tu",
-      "phone": "123123123123",
-      "id": "lyPEyrs2l"
-    },
-    {
-      "name": "asdtutu",
-      "phone": "21312313213",
-      "id": "8gMcIGbB6"
     }
   ]
 }
