+++ routes/user.routes.js
@@ -1,6 +1,6 @@
 const express = require('express')
 const router = express.Router()
-
+const validate = require('../validate/user.validate')
 
 const controller = require('../controllers/user.controller')
 
@@ -13,7 +13,7 @@ router.get('/create',controller.create)
 
 router.get('/:userId',controller.get )
 
-router.post('/create',controller.postCreate)
+router.post('/create',validate.postCreate ,controller.postCreate)
